cli kubernetes mode
