+++ airflow/bin/cli/__init__.py
@@ -1,0 +1,3 @@
+from cli import *
+from cli_factory import *
+from airflow import *
